Fix app crash when user cancel first student creation
