+++ src/screens/studentSettings/StudentSettingsScreen.tsx
@@ -1,13 +1,83 @@
 import React from 'react';
+import { StyleSheet, View } from 'react-native';
 import { NavigationInjectedProps } from 'react-navigation';
 
-import CreateStudent from './CreateStudent';
-import StudentSettings from './StudentSettings';
+import { FlatButton, NarrowScreenTemplate } from 'components';
+import { i18n } from 'locale';
+import { ModelSubscriber, Student } from 'models';
+import { palette } from 'styles';
+import { StudentPanel } from './StudentPanel';
 
-export class StudentSettingsScreen extends React.PureComponent<NavigationInjectedProps> {
+interface State {
+  student: Student;
+}
+
+export class StudentSettingsScreen extends React.PureComponent<NavigationInjectedProps, State> {
+  studentSubscriber: ModelSubscriber<Student> = new ModelSubscriber();
+
+  constructor(props: NavigationInjectedProps) {
+    super(props);
+    this.state = {
+      student: props.navigation.getParam('student'),
+    };
+  }
+
+  componentDidMount() {
+    const student = this.props.navigation.getParam('student');
+    this.studentSubscriber.subscribeElementUpdates(student, updatedStudent =>
+      this.setState({ student: updatedStudent }),
+    );
+  }
+
+  componentWillUnmount() {
+    this.studentSubscriber.unsubscribeElementUpdates();
+  }
+
+  get screenName(): string {
+    return i18n.t('studentSettings:settingsTitle', {
+      studentName: this.props.navigation.getParam('student').name,
+    });
+  }
+
+  handleChange = (name: string) => this.setState({ student: { ...this.state.student, name } });
+
+  handleEndEditing = () => {
+    const { student } = this.state;
+
+    if (!student.name) {
+      return;
+    }
+
+    this.state.student.update({ name: student.name });
+  };
+
+  handleRemoveStudent = () => this.state.student.delete().then(() => this.props.navigation.goBack());
+
   render() {
-    const createStudent = this.props.navigation.getParam('createStudent');
+    const { student } = this.state;
 
-    return <>{createStudent ? <CreateStudent /> : <StudentSettings />}</>;
+    return (
+      <NarrowScreenTemplate title={this.screenName} navigation={this.props.navigation}>
+        <StudentPanel
+          student={student}
+          handleChangeName={this.handleChange}
+          handleEndEditingName={this.handleEndEditing}
+        >
+          <View>
+            <FlatButton
+              title={i18n.t('studentSettings:removeStudent')}
+              titleStyle={styles.studentButton}
+              onPress={this.handleRemoveStudent}
+            />
+          </View>
+        </StudentPanel>
+      </NarrowScreenTemplate>
+    );
   }
 }
+
+const styles = StyleSheet.create({
+  studentButton: {
+    color: palette.textBody,
+  },
+});
